+++ CHANGES.rst
@@ -1,3 +1,19 @@
+0.9.3 (unreleased)
+------------------
+
+- No changes yet
+
+0.9.2 (2014-12-01)
+------------------
+
+- Added support for a new model directory format that stores all SEDs in a
+  single file. [#16]
+
+- Added a new Fitter class that provides an OO interface to the fitter. This
+  makes it easier to fit sources one by one without reloading models. [#12]
+
+- Various bug fixes.
+
 0.9.1 (2014-03-10)
 ------------------